output.rs: Extract some fixtures from tests
